requirements.txt
streamlit
requests
PyGithub